requirements.txt
flask
simplejson
PyAudio
pyzmq
